+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.6] 2020-01-19
+### Added
+- Simple sync of votes, may drop some votes
+
+### Fixed
+- Broken stations were not marked
+- Make foreign keys of StationClick and StationCheckHistory on delete cascade
+
 ## [0.6.5] 2020-01-18
 ### Added
 - Ansible role and example playbook for debian/ubuntu